+++ girder_annotation/girder_large_image_annotation/web_client/annotations/defaults/circle.js
@@ -1,0 +1,5 @@
+export default {
+    fillColor: 'rgba(0,0,0,0)',
+    lineColor: 'rgb(0,0,0)',
+    lineWidth: 2
+};