+++ girder_annotation/girder_large_image_annotation/web_client/annotations/defaults/ellipse.js
@@ -1,0 +1,7 @@
+export default {
+    fillColor: 'rgba(0,0,0,0)',
+    lineColor: 'rgb(0,0,0)',
+    lineWidth: 2,
+    rotation: 0,
+    normal: [0, 0, 1]
+};