+++ girder_annotation/girder_large_image_annotation/web_client/annotations/geojs/types/index.js
@@ -1,11 +1,15 @@
 import point from './point';
 import rectangle from './rectangle';
+import ellipse from './ellipse';
+import circle from './circle';
 import polygon from './polygon';
 import line from './line';
 
 export {
     point,
     rectangle,
+    ellipse,
+    circle,
     polygon,
     line
 };