+++ girder_annotation/girder_large_image_annotation/web_client/annotations/geometry/index.js
@@ -1,9 +1,13 @@
 import rectangle from './rectangle';
+import ellipse from './ellipse';
+import circle from './circle';
 import polyline from './polyline';
 import point from './point';
 
 export {
     rectangle,
+    ellipse,
+    circle,
     polyline,
     point
 };